+++ plugin/src/test/java/com/stratio/cassandra/lucene/schema/mapping/DoubleMapperTest.java
@@ -129,7 +129,6 @@ public class DoubleMapperTest extends AbstractMapperTest {
         DoubleMapper mapper = doubleMapper().boost(1f).build("field");
         Double parsed = mapper.base("test", 3.5f);
         assertEquals("Base for floats is wrong", Double.valueOf(3.5d), parsed);
-
     }
 
     @Test
@@ -151,7 +150,6 @@ public class DoubleMapperTest extends AbstractMapperTest {
         DoubleMapper mapper = doubleMapper().boost(1f).build("field");
         Double parsed = mapper.base("test", 3.5d);
         assertEquals("Base for doubles is wrong", Double.valueOf(3.5d), parsed);
-
     }
 
     @Test
@@ -159,7 +157,6 @@ public class DoubleMapperTest extends AbstractMapperTest {
         DoubleMapper mapper = doubleMapper().boost(1f).build("field");
         Double parsed = mapper.base("test", 3.6d);
         assertEquals("Base for doubles is wrong", Double.valueOf(3.6d), parsed);
-
     }
 
     @Test